[auth-svc] Redis repositories
