+++ services/auth-svc/go.mod
@@ -3,9 +3,11 @@ module github.com/hausops/mono/services/auth-svc
 go 1.20
 
 require (
+	github.com/alicebob/miniredis/v2 v2.30.3
 	github.com/brianvoe/gofakeit v3.18.0+incompatible
 	github.com/grpc-ecosystem/go-grpc-middleware v1.4.0
 	github.com/hausops/mono/services/user-svc v0.0.0-20230529135640-7edd4954033f
+	github.com/redis/go-redis/v9 v9.0.5
 	github.com/rs/xid v1.5.0
 	github.com/wagslane/go-password-validator v0.3.0
 	go.uber.org/zap v1.24.0
@@ -17,11 +19,15 @@ require (
 )
 
 require (
+	github.com/alicebob/gopher-json v0.0.0-20200520072559-a9ecdc9d1d3a // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
 	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/yuin/gopher-lua v1.1.0 // indirect
 	go.uber.org/atomic v1.7.0 // indirect
 	go.uber.org/multierr v1.6.0 // indirect
-	golang.org/x/net v0.8.0 // indirect
-	golang.org/x/sys v0.6.0 // indirect
-	golang.org/x/text v0.8.0 // indirect
+	golang.org/x/net v0.10.0 // indirect
+	golang.org/x/sys v0.8.0 // indirect
+	golang.org/x/text v0.9.0 // indirect
 	google.golang.org/genproto v0.0.0-20230306155012-7f2fa6fef1f4 // indirect
 )
